Fix #2 (Position desync) but breaks ... client prediction !
At least, it seems like that the server doesn't simulate accurately the
movements.
Thus, causing the local player to think that it should be there whereas it
shouldn't.

Also, when spamming and doing a lot of stress movement, you'll suffer from
excessive lag and teleport.

Seems like this issue isn't a position desync, but a server bad
simulation.